+++ src/combinator/decorator/flow.rs
@@ -9,13 +9,13 @@ create_closure_decorator!(Reject, "See [`Combinator::reject`].");
 create_simple_decorator!(Optional, "See [`Combinator::optional`].");
 create_simple_decorator!(Boundary, "See [`Combinator::boundary`].");
 
-unsafe impl<State, Heap, T: Action<State, Heap>, D: Fn(Input<&mut State, &mut Heap>) -> bool>
+unsafe impl<State, Heap, T: Action<State, Heap>, D: Fn(Input<&str, &mut State, &mut Heap>) -> bool>
   Action<State, Heap> for When<T, D>
 {
   type Value = T::Value;
 
   #[inline]
-  fn exec(&self, mut input: Input<&mut State, &mut Heap>) -> Option<Output<Self::Value>> {
+  fn exec(&self, mut input: Input<&str, &mut State, &mut Heap>) -> Option<Output<Self::Value>> {
     if (self.inner)(input.reborrow()) {
       self.action.exec(input)
     } else {
@@ -24,13 +24,13 @@ unsafe impl<State, Heap, T: Action<State, Heap>, D: Fn(Input<&mut State, &mut He
   }
 }
 
-unsafe impl<State, Heap, T: Action<State, Heap>, D: Fn(Input<&mut State, &mut Heap>) -> bool>
+unsafe impl<State, Heap, T: Action<State, Heap>, D: Fn(Input<&str, &mut State, &mut Heap>) -> bool>
   Action<State, Heap> for Prevent<T, D>
 {
   type Value = T::Value;
 
   #[inline]
-  fn exec(&self, mut input: Input<&mut State, &mut Heap>) -> Option<Output<Self::Value>> {
+  fn exec(&self, mut input: Input<&str, &mut State, &mut Heap>) -> Option<Output<Self::Value>> {
     if !(self.inner)(input.reborrow()) {
       self.action.exec(input)
     } else {
@@ -43,13 +43,13 @@ unsafe impl<
     State,
     Heap,
     T: Action<State, Heap>,
-    D: Fn(AcceptedContext<Input<&mut State, &mut Heap>, &Output<T::Value>>) -> bool,
+    D: Fn(AcceptedContext<Input<&str, &mut State, &mut Heap>, &Output<T::Value>>) -> bool,
   > Action<State, Heap> for Reject<T, D>
 {
   type Value = T::Value;
 
   #[inline]
-  fn exec(&self, mut input: Input<&mut State, &mut Heap>) -> Option<Output<Self::Value>> {
+  fn exec(&self, mut input: Input<&str, &mut State, &mut Heap>) -> Option<Output<Self::Value>> {
     self.action.exec(input.reborrow()).and_then(|output| {
       if (self.inner)(AcceptedContext {
         input,
@@ -69,7 +69,7 @@ unsafe impl<State, Heap, T: Action<State, Heap, Value: Default>> Action<State, H
   type Value = T::Value;
 
   #[inline]
-  fn exec(&self, input: Input<&mut State, &mut Heap>) -> Option<Output<Self::Value>> {
+  fn exec(&self, input: Input<&str, &mut State, &mut Heap>) -> Option<Output<Self::Value>> {
     Some(self.action.exec(input).unwrap_or_else(|| Output {
       value: Default::default(),
       digested: 0,
@@ -81,7 +81,7 @@ unsafe impl<State, Heap, T: Action<State, Heap>> Action<State, Heap> for Boundar
   type Value = T::Value;
 
   #[inline]
-  fn exec(&self, mut input: Input<&mut State, &mut Heap>) -> Option<Output<Self::Value>> {
+  fn exec(&self, mut input: Input<&str, &mut State, &mut Heap>) -> Option<Output<Self::Value>> {
     self.action.exec(input.reborrow()).and_then(|output| {
       unsafe { input.instant().rest().get_unchecked(output.digested..) }
         .chars()
@@ -106,7 +106,7 @@ impl<T> Combinator<T> {
   /// # ;}
   /// ```
   #[inline]
-  pub fn when<State, Heap, F: Fn(Input<&mut State, &mut Heap>) -> bool>(
+  pub fn when<State, Heap, F: Fn(Input<&str, &mut State, &mut Heap>) -> bool>(
     self,
     condition: F,
   ) -> Combinator<When<T, F>>
@@ -129,7 +129,7 @@ impl<T> Combinator<T> {
   /// # ;}
   /// ```
   #[inline]
-  pub fn prevent<State, Heap, F: Fn(Input<&mut State, &mut Heap>) -> bool>(
+  pub fn prevent<State, Heap, F: Fn(Input<&str, &mut State, &mut Heap>) -> bool>(
     self,
     preventer: F,
   ) -> Combinator<Prevent<T, F>>
@@ -152,7 +152,7 @@ impl<T> Combinator<T> {
   pub fn reject<
     State,
     Heap,
-    F: Fn(AcceptedContext<Input<&mut State, &mut Heap>, &Output<T::Value>>) -> bool,
+    F: Fn(AcceptedContext<Input<&str, &mut State, &mut Heap>, &Output<T::Value>>) -> bool,
   >(
     self,
     rejecter: F,