+++ src/combinator/decorator/state.rs
@@ -6,13 +6,13 @@ create_closure_decorator!(Then, "See [`Combinator::then`].");
 create_closure_decorator!(Catch, "See [`Combinator::catch`].");
 create_closure_decorator!(Finally, "See [`Combinator::finally`].");
 
-unsafe impl<State, Heap, T: Action<State, Heap>, D: Fn(Input<&mut State, &mut Heap>)>
+unsafe impl<State, Heap, T: Action<State, Heap>, D: Fn(Input<&str, &mut State, &mut Heap>)>
   Action<State, Heap> for Prepare<T, D>
 {
   type Value = T::Value;
 
   #[inline]
-  fn exec(&self, mut input: Input<&mut State, &mut Heap>) -> Option<Output<Self::Value>> {
+  fn exec(&self, mut input: Input<&str, &mut State, &mut Heap>) -> Option<Output<Self::Value>> {
     (self.inner)(input.reborrow());
     self.action.exec(input)
   }
@@ -22,26 +22,26 @@ unsafe impl<
     State,
     Heap,
     T: Action<State, Heap>,
-    D: Fn(AcceptedContext<Input<&mut State, &mut Heap>, &Output<T::Value>>),
+    D: Fn(AcceptedContext<Input<&str, &mut State, &mut Heap>, &Output<T::Value>>),
   > Action<State, Heap> for Then<T, D>
 {
   type Value = T::Value;
 
   #[inline]
-  fn exec(&self, mut input: Input<&mut State, &mut Heap>) -> Option<Output<Self::Value>> {
+  fn exec(&self, mut input: Input<&str, &mut State, &mut Heap>) -> Option<Output<Self::Value>> {
     self.action.exec(input.reborrow()).inspect(|output| {
       (self.inner)(AcceptedContext { input, output });
     })
   }
 }
 
-unsafe impl<State, Heap, T: Action<State, Heap>, D: Fn(Input<&mut State, &mut Heap>)>
+unsafe impl<State, Heap, T: Action<State, Heap>, D: Fn(Input<&str, &mut State, &mut Heap>)>
   Action<State, Heap> for Catch<T, D>
 {
   type Value = T::Value;
 
   #[inline]
-  fn exec(&self, mut input: Input<&mut State, &mut Heap>) -> Option<Output<Self::Value>> {
+  fn exec(&self, mut input: Input<&str, &mut State, &mut Heap>) -> Option<Output<Self::Value>> {
     let output = self.action.exec(input.reborrow());
     if output.is_none() {
       (self.inner)(input);
@@ -50,13 +50,13 @@ unsafe impl<State, Heap, T: Action<State, Heap>, D: Fn(Input<&mut State, &mut He
   }
 }
 
-unsafe impl<State, Heap, T: Action<State, Heap>, D: Fn(Input<&mut State, &mut Heap>)>
+unsafe impl<State, Heap, T: Action<State, Heap>, D: Fn(Input<&str, &mut State, &mut Heap>)>
   Action<State, Heap> for Finally<T, D>
 {
   type Value = T::Value;
 
   #[inline]
-  fn exec(&self, mut input: Input<&mut State, &mut Heap>) -> Option<Output<Self::Value>> {
+  fn exec(&self, mut input: Input<&str, &mut State, &mut Heap>) -> Option<Output<Self::Value>> {
     let output = self.action.exec(input.reborrow());
     (self.inner)(input);
     output
@@ -75,7 +75,7 @@ impl<T> Combinator<T> {
   /// # ;}
   /// ```
   #[inline]
-  pub fn prepare<State, Heap, F: Fn(Input<&mut State, &mut Heap>)>(
+  pub fn prepare<State, Heap, F: Fn(Input<&str, &mut State, &mut Heap>)>(
     self,
     modifier: F,
   ) -> Combinator<Prepare<T, F>>
@@ -99,7 +99,7 @@ impl<T> Combinator<T> {
   pub fn then<
     State,
     Heap,
-    F: Fn(AcceptedContext<Input<&mut State, &mut Heap>, &Output<T::Value>>),
+    F: Fn(AcceptedContext<Input<&str, &mut State, &mut Heap>, &Output<T::Value>>),
   >(
     self,
     modifier: F,
@@ -121,7 +121,7 @@ impl<T> Combinator<T> {
   /// # ;}
   /// ```
   #[inline]
-  pub fn catch<State, Heap, F: Fn(Input<&mut State, &mut Heap>)>(
+  pub fn catch<State, Heap, F: Fn(Input<&str, &mut State, &mut Heap>)>(
     self,
     modifier: F,
   ) -> Combinator<Catch<T, F>>
@@ -143,7 +143,7 @@ impl<T> Combinator<T> {
   /// # ;}
   /// ```
   #[inline]
-  pub fn finally<State, Heap, F: Fn(Input<&mut State, &mut Heap>)>(
+  pub fn finally<State, Heap, F: Fn(Input<&str, &mut State, &mut Heap>)>(
     self,
     modifier: F,
   ) -> Combinator<Finally<T, F>>
@@ -166,14 +166,14 @@ mod tests {
   }
 
   fn accepter() -> Combinator<impl Action<State, Value = ()>> {
-    wrap(|input: Input<&mut State, &mut ()>| {
+    wrap(|input: Input<&str, &mut State, &mut ()>| {
       input.state.to = input.state.from;
       input.digest(1)
     })
   }
 
   fn rejecter() -> Combinator<impl Action<State, Value = ()>> {
-    wrap(|input: Input<&mut State, &mut ()>| {
+    wrap(|input: Input<&str, &mut State, &mut ()>| {
       input.state.to = input.state.from;
       None
     })