+++ src/combinator/ops/bitor.rs
@@ -54,7 +54,7 @@ unsafe impl<State, Heap, Lhs: Action<State, Heap>, Rhs: Action<State, Heap, Valu
   type Value = Lhs::Value;
 
   #[inline]
-  fn exec(&self, mut input: Input<&mut State, &mut Heap>) -> Option<Output<Self::Value>> {
+  fn exec(&self, mut input: Input<&str, &mut State, &mut Heap>) -> Option<Output<Self::Value>> {
     self
       .lhs
       .exec(input.reborrow())