+++ src/general.rs
@@ -1,5 +1,0 @@
-mod action;
-mod instant;
-
-pub use action::*;
-pub use instant::*;