+++ src/lib.rs
@@ -5,5 +5,3 @@ pub mod combinator;
 pub mod instant;
 pub mod parser;
 pub mod range;
-
-pub mod general;